==> 05_targeted_metaB_BAs.Rmd <==
---
title: "continuous culture - targeted metaB - BAs"
output: html_document
---

#<!-- Load required packages -->
```{r}
library(data.table)
library(ggplot2)
library(gtools)
library(reshape2)
```

<!-- Define paths -->
```{r}
datapath <- "data/continuous_culture/"
```

<!-- Load data -->
```{r}
ts4_file<- paste0(datapath, 'TableS4_metaB_target_bile_acids.csv')
bas_metab_metadata <- as.data.frame(fread(paste0(datapath, 'TableS4_metaB_target_bile_acids.csv'),  header = T), stringsAsFactors = F)

# Sample IDs
bas_metab_samples <- as.data.frame(t(bas_metab_metadata[c(1:2),]))
bas_metab_samples$header <- rownames(bas_metab_samples)
names(bas_metab_samples) <- bas_metab_samples[1,]
bas_metab_samples <- bas_metab_samples[-c(1),] 

# BAs concentrations
bas_metab_conc <- (bas_metab_metadata[-c(1:2),])
bas_metab_conc_melt <- melt(bas_metab_conc, id.vars=c("header"))
names(bas_metab_conc_melt) <- c('bas', 'header', 'value')
bas_metab_conc_melt2 <- merge(bas_metab_conc_melt, bas_metab_samples, by='header')
bas_metab_conc_melt2$value <- as.numeric(bas_metab_conc_melt2$value)
bas_metab_conc_melt2$time <- as.numeric(bas_metab_conc_melt2$time)
```


<!-- Secondary - DCA and LCA - fMolar -->
# Figure 2. C. scindens-mediated biotransformation of primary BAs and downregulation of C. difficile toxin tcdA expression.
## C) Concentration of DCA and LCA in the treatment condition compared to the control condition in a continuous culture setup, using targeted metabolomics. 
```{r}
# Subset data by secondaria BAs DCA and LCA
bas_metab_metadata_var <- bas_metab_conc_melt2[bas_metab_conc_melt2$bas %in% c("DCA","LCA"),]

bas_metab_metadata_var$value[bas_metab_metadata_var$value == 0 & bas_metab_metadata_var$Condition == 'Cdiff'] <- NA

# transform concentration: nanoMolar
# BAs     ng/L       nM
# DCA 	392.6	       1
# LCA  	376.6	       1

bas_metab_metadata_var$value_nM <- (bas_metab_metadata_var$value)/392.6
bas_metab_metadata_var$value_nM[bas_metab_metadata_var$variable == 'LCA'] <- (bas_metab_metadata_var$value[bas_metab_metadata_var$variable == 'LCA'])/376.6

colors_kit =c("Cdiff" = "#92392F", "Cdiff_Csind" = "#11698E") 
dca_lca_plot <- ggplot(data=subset(bas_metab_metadata_var, bas %in% c("DCA","LCA")), 
                     aes(x=time, y=value_nM, color = Condition)) +
  geom_point(alpha=1) + 
  #geom_bar(aes(y=RA_sum, fill = taxa_ID),stat="identity", alpha=.7) +
  theme_bw() + labs(y = 'Concentrration (nM)', x = 'Time (hours)') + 
  theme(axis.text = element_text(size = 10),
        axis.text.x = element_text(color = "grey20", size = 9,angle = 0, face = "plain"),
        axis.text.y = element_text(color = "grey20", size = 10, angle = 0, hjust = 1, vjust = 0, face = "plain"),
        axis.title.x = element_text(color = "grey20", size = 12, angle = 0, hjust = 0.5, vjust = 1, face = "plain"),
        axis.title.y = element_text(color = "grey20", size = 12, angle = 90, hjust = .5, vjust = .5, face = "plain"), 
        panel.grid.major.x = element_blank(), panel.grid.minor = element_blank(), legend.position="bottom") + 
  guides(color=guide_legend(nrow= 1))+ 
  facet_wrap(.~bas, scales = 'free')+
  scale_color_manual(values = colors_kit, name="Condition")

print(dca_lca_plot + geom_line(alpha = 1,size = 0.65))


#####
# effect size - DCA
bas_metab_metadata_var$value[is.na(bas_metab_metadata_var$value)] <- 0
bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff' & 
                                           bas_metab_metadata_var$time=='25' & 
                                           bas_metab_metadata_var$bas == "DCA")]/
  bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff' & 
                                             bas_metab_metadata_var$time=='1' & 
                                             bas_metab_metadata_var$bas == "DCA")] # 0.6907626
## Cdiff_Csind
bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff_Csind' & 
                                           bas_metab_metadata_var$time=='25' & 
                                           bas_metab_metadata_var$bas == "DCA")]/
  bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff_Csind' & 
                                             bas_metab_metadata_var$time=='1' & 
                                             bas_metab_metadata_var$bas == "DCA")] # 10.71193

# effect size - LCA
bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff' & 
                                           bas_metab_metadata_var$time=='25' & 
                                           bas_metab_metadata_var$bas == "LCA")]/
  bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff' & 
                                             bas_metab_metadata_var$time=='1' & 
                                             bas_metab_metadata_var$bas == "LCA")] # NA
## Cdiff_Csind
bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff_Csind' & 
                                           bas_metab_metadata_var$time=='25' & 
                                           bas_metab_metadata_var$bas == "LCA")]/
  bas_metab_metadata_var$value[(bas_metab_metadata_var$Condition=='Cdiff_Csind' & 
                                             bas_metab_metadata_var$time=='1' & 
                                             bas_metab_metadata_var$bas == "LCA")] # Inf


```
